pyproject.toml
[project]
name = "gerador-historia"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "requests"
]